small fix
git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@598 70169cfe-8b10-0410-8925-dcb4b91034d8

+++ src/parser/csv.c
@@ -74,7 +74,8 @@ int GWEN_CSV_GetNameAndIndex(const char *name,
       j++;
     } /* while */
     if (j>=sizeof(numbuffer)) {
-      DBG_INFO(GWEN_LOGDOMAIN, "Index number too long (%u>=%u)", j, sizeof(numbuffer));
+      DBG_INFO(GWEN_LOGDOMAIN, "Index number too long (%u>=%d)", j,
+               (int)(sizeof(numbuffer)));
       return -1;
     }
     numbuffer[j]=0;
